refactor: [#184] convert test command handler to folder module
- Convert single-file module to folder structure
- Separate handler logic, errors, and tests into dedicated files
- Update execute signature to accept &EnvironmentName
- Add repository field to TestCommandHandler
- Align with provision and configure handler patterns
- Update E2E task to use new handler signature
- All tests pass and pre-commit checks successful
