+++ src/application/command_handlers/test/tests/mod.rs
@@ -1,0 +1,8 @@
+//! Tests for test command handler
+
+mod integration;
+
+#[cfg(test)]
+pub(crate) mod builders {
+    // Test builders if needed in the future
+}
